Don't install phing from PEAR on travis - we have it in composer
The composer.json was updated in daf1e3d to install phing as a composer
package. It's therefore already on our travis instance (in vendor/bin)
without also needing to install from PEAR.

This resolves build errors on hhvm (where pear is not installed by
default) and reduces build time on other environments.